move window-frame-updater
